+++ src/test/java/net/snowflake/client/jdbc/BindUploaderLatestIT.java
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2012-2020 Snowflake Computing Inc. All right reserved.
+ */
 package net.snowflake.client.jdbc;
 
 import static net.snowflake.client.jdbc.BindUploaderIT.*;
@@ -23,6 +26,12 @@ import net.snowflake.client.core.bind.BindUploader;
 import org.junit.*;
 import org.junit.experimental.categories.Category;
 
+/**
+ * Bind Uploader tests for the latest JDBC driver. This doesn't work for the oldest supported
+ * driver. Revisit this tests whenever bumping up the oldest supported driver to examine if the
+ * tests still is not applicable. If it is applicable, move tests to BindUploaderIT so that both the
+ * latest and oldest supported driver run the tests.
+ */
 @Category(TestCategoryOthers.class)
 public class BindUploaderLatestIT extends BaseJDBCTest {
   BindUploader bindUploader;
