+++ src/test/java/net/snowflake/client/jdbc/PreparedStatement2LatestIT.java
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2012-2020 Snowflake Computing Inc. All right reserved.
+ */
 package net.snowflake.client.jdbc;
 
 import static org.hamcrest.CoreMatchers.is;
@@ -13,10 +16,10 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
 /**
- * PreparedStatement integration tests for the latest JDBC driver but not the oldest supported
- * driver. Revisit this tests whenever bumping up the oldest supported driver to examine if the
- * tests still is not applicable. If it is applicable, move tests to PreparedStatement2IT so that
- * both the latest and oldest supported driver run the tests.
+ * PreparedStatement integration tests for the latest JDBC driver. This doesn't work for the oldest
+ * supported driver. Revisit this tests whenever bumping up the oldest supported driver to examine
+ * if the tests still are not applicable. If it is applicable, move tests to PreparedStatement2IT so
+ * that both the latest and oldest supported driver run the tests.
  */
 @Category(TestCategoryStatement.class)
 public class PreparedStatement2LatestIT extends PreparedStatement0IT {
