+++ src/test/java/net/snowflake/client/jdbc/PreparedStatementArrow2IT.java
@@ -8,7 +8,7 @@ import org.junit.experimental.categories.Category;
 
 /** Test PreparedStatement in ARROW format 2/2 */
 @Category(TestCategoryArrow.class)
-public class PreparedStatementArrow2IT extends PreparedStatement2IT {
+public class PreparedStatementArrow2IT extends PreparedStatement2LatestIT {
   public PreparedStatementArrow2IT() {
     super("arrow");
   }
