+++ src/test/java/net/snowflake/client/jdbc/StreamLatestIT.java
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2012-2020 Snowflake Computing Inc. All right reserved.
+ */
 package net.snowflake.client.jdbc;
 
 import static org.junit.Assert.assertEquals;
@@ -10,6 +13,12 @@ import net.snowflake.client.category.TestCategoryOthers;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
+/**
+ * Stream API tests for the latest JDBC driver. This doesn't work for the oldest supported driver.
+ * Revisit this tests whenever bumping up the oldest supported driver to examine if the tests still
+ * is not applicable. If it is applicable, move tests to StreamIT so that both the latest and oldest
+ * supported driver run the tests.
+ */
 @Category(TestCategoryOthers.class)
 public class StreamLatestIT extends BaseJDBCTest {
   /**
